requirements.txt
-e .
aws-cdk.aws-s3
aws_cdk.aws_iam